Feat: Enable Use of MinGW-w64's GCC in Windows
All MinGW-gcc tests build and pass (when run manually) with this
version, so long as we manually create a softlink to event_core.dll
from the tests directory.

meson.build
Make multiple Windows libraries link statically (may undo this later)

src/meson.build
Updated some comments
In MinGW-gcc case, use winres.exe (MinGW resource compiler) not rc.exe
For MinGW-gcc, don't create static lib, generate .def file etc. - not
needed
For MinGW-gcc, no '/noentry' option needed for linker

tests/meson.build
Add additionl softlink to pistache.dll for benefit of test programs
In MinGW-gcc case, link std c++ libs statically, as recommended
(avoids Windows accidentally loading the Microsoft ones).

async.h
Revert to use "auto& req" rather tha "auto req", where req is a
pointer to a request. MinGW-gcc objects warns on additional pointer,
prefers reference to pointer.

pist_timelog.h
reinterpret_cast pointers to "malloc" and "free" functions to void *,
to avoid MinGW-gcc warning.

stream.h
Explicitly set locale of std::basic_streambuf to avoid intermittent
crash in std::basic_streambuf destructor. Possible MinGW-gcc bug.

winornix.h
Only define NOMINMAX if not already defined
Provide PST_STRERROR_R for MinGW-gcc case
Correct parameter casts in sock-read, sock-write, file-read and
file-write

pist_clock_gettime.cc
Use standard-compliant integer-type-annotation (e.g. suffix with 'll'
not 'i64').

pist_filefns.cc
Add header file for, and use, _SH_DENYNO even in MinGW

pist_ifaddrs.cc
Use 'delete[]' where necessary, not just 'delete'.

pist_resource.cc
Do static_cast as needed for tv_usec value

pist_strerror_r.cc
Remove EOTHER (which doesn't exist for MinGW) from switch statement
for MinGW.

pist_syslog.cc
Manually define _Pre_cap_, macro needed by the generated header file
pist_winlog.h.
Suppress unknown pragma warnings from pist_winlog.h
Manually supply an instance of Pistache_ProviderEnableBits for
pist_winlog.h
join() each std::thread before allowing its destructor to complete by
creating a new lLogToStdOutAsWellInst whose destructor carries out the
joins.

listener.cc
Suppress 'unused-value' warning when using SSL macro

router.cc
Make array of char 1 char big, not 0, to avoid warning about creation
of shared_ptr to the array.

http_server_shutdown.cc
Don't explicitly include ConsoleApi.h, not needed and confuses
MinGW-gcc.

dump2def.cc
use 'int' for argc type

winscripts/adjbuilddirformesbuild.ps1
Automatically add '.gcc' to build dir name if gcc is to be used

winscripts/mesbuild.ps1
winscripts/mesbuilddebug.ps1
Make verbose for now

+++ src/common/pist_resource.cc
@@ -56,12 +56,12 @@ extern "C" int pist_getrusage(int who, struct PST_RUSAGE * rusage)
     memcpy(&li, &kerneltime, sizeof(FILETIME));
     li.QuadPart /= 10L;            /* Convert to microseconds */
     rusage->ru_stime.tv_sec = (long) (li.QuadPart / 1000000L);
-    rusage->ru_stime.tv_usec = li.QuadPart % 1000000L;
+    rusage->ru_stime.tv_usec = static_cast<long int>(li.QuadPart % 1000000L);
 
     memcpy(&li, &usertime, sizeof(FILETIME));
     li.QuadPart /= 10L;            /* Convert to microseconds */
     rusage->ru_utime.tv_sec = (long) (li.QuadPart / 1000000L);
-    rusage->ru_utime.tv_usec = li.QuadPart % 1000000L;
+    rusage->ru_utime.tv_usec = static_cast<long int>(li.QuadPart % 1000000L);
 
     return(0); // success
 }
